Compacted the level file

+++ Informatik/Informatik/entity/Player.hpp
@@ -32,7 +32,7 @@ private:
     
 public:
     float _x = 0, _y = 0; // Coordinates
-    int _z = 0;// Height level
+    uint8_t _z = 0;// Height level
     bool inControl = true; // Player can control himself by default
